+++ src/server/queries/event.ts
@@ -72,7 +72,7 @@ export interface OrganizerViewEvent {
   }[]
   invitations: {
     email: string
-    status: "pending" | "expired"
+    status: "pending" | "expired" | "rejected"
   }[]
   exclusionRules: {
     giver: {
@@ -149,7 +149,7 @@ export async function getEventInfo(eventId: string): Promise<EventViewReturn> {
   const [participants, invitations, exclusionRules, assignments] =
     await Promise.all([
       fetchParticipants(eventId),
-      fetchPendingInvitations(eventId),
+      fetchNotAcceptedInvitations(eventId),
       fetchExclusionRules(eventId),
       fetchAssignments(eventId),
     ])
@@ -188,20 +188,20 @@ function fetchParticipants(eventId: string) {
     .where(eq(eventParticipant.eventId, eventId))
 }
 
-// return query to fetch pending or expired invitations of an event
-function fetchPendingInvitations(eventId: string) {
+// return query to fetch expired invitations of an event
+function fetchNotAcceptedInvitations(eventId: string) {
   return db
     .select({
       email: invitation.email,
       status: sql<
-        "pending" | "expired"
-      >`CASE WHEN ${invitation.expiresAt} > now() THEN 'pending' ELSE 'expired' END`,
+        "pending" | "expired" | "rejected"
+      >`CASE WHEN ${invitation.expiresAt} < now() THEN 'expired' ELSE ${invitation.status} END`,
     })
     .from(invitation)
     .where(
       and(
         eq(invitation.eventId, eventId),
-        eq(invitation.accepted, false),
+        ne(invitation.status, "accepted"),
         eq(invitation.revoked, false)
       )
     )
